% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse}
\alias{analyse}
\title{analyse data with 7 different methods (covLASSO, covRidge, covShrink, semLASSO, semRidge, regular ML sem, regular FIML sem)}
\usage{
analyse(dat)
}
\arguments{
\item{dat}{data set}
}
\description{
analyse data with 7 different methods (covLASSO, covRidge, covShrink, semLASSO, semRidge, regular ML sem, regular FIML sem)
}
