% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulateMIMIC}
\alias{simulateMIMIC}
\title{simulate data for a MIMIC model}
\usage{
simulateMIMIC(p, N, seed)
}
\arguments{
\item{p}{number of predictors (currently with fixed collinearities: 2/5 = 0, 1/5 = 0.2, 0.5 and 0.8 respectively)}

\item{N}{sample size}

\item{seed}{seed for random number generation}
}
\description{
simulate data for a MIMIC model
}
